requirements.txt
# API依赖
openai>=1.0.0

# 数据处理
pandas>=1.5.0

# 工具依赖
tqdm>=4.60.0
python-dotenv>=0.19.0

# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.7.0

# 代码格式化
black>=23.3.0
flake8>=6.0.0
